requirements.txt
asgiref==3.5.1
black==22.3.0
click==8.1.3
Django==3.2.13
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
importlib-metadata==4.11.3
mypy-extensions==0.4.3
oauth2-provider==0.0
pathspec==0.9.0
platformdirs==2.5.2
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typed-ast==1.5.3
typing_extensions==4.2.0
zipp==3.8.0
